refactor(cosim): split cosim_metal.rs into cosim/{mod,metal}.rs (P1.4)
mod.rs (non-gated): CosimBackend trait, BitOp/FlashDebug, CosimOpts/Result,
run_cosim_generic<B>, patchers, scheduler + agnostic glue. metal.rs
(#[cfg(feature="metal")]): MetalBackend/MetalSimulator/ScheduleBuffers, GPU
structs, build_*/encode_*, the run_cosim shim. src/sim/mod.rs now `pub mod
cosim;` (non-gated); jacquard.rs paths updated. cargo check --lib
--no-default-features now compiles the agnostic half. Metal bit-identical,
298 tests pass.

Co-developed-by: Claude Code v2.1.168 (claude-opus-4-8)
